# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Berim API",
    "description": "<div style='direction: rtl'>به کمک نرم‌افزار بریم با دوستان خود برنامه بیرون رفتن هماهنگ کنید. <br/> این سند از چهار بخش تشکیل شده که در هر بخش توابع مربوط به همان بخش به همراه ورودی ها و خروجی های آن امده است. این بخش ها عبارتند از کاربر، مکان‌ها، دوستان و گفتگوها. سعی شده توضیحات هر تابع درون خود تابع داده شود و سند گویا باشد. به کمک شما دوستان کم کم این سند رو تکمیل خواهیم کرد. <br/> از دیگر مزیت های این سیستم قابلیت تست ای پی ای هست که در توابع زیر خواهید دید که میتونیم ورودی های هر تابع رو وارد کنیم و نتیجه خروجی رو ببینیم که مطمئنا خیلی کمکمون خواهد کرد. <br/> با سپاس فراوان :)</div>",
    "version": "1.0.0"
  },
  "produces": ["application/json"],
  "host": "localhost:8080",
  "basePath": "/v1",
  "paths": {
...
  "/friend/answer_friend_request": {
    
    "post": {
      "summary": "answer friend request",
      "description":"پاسخ به درخواست دوستی",
      "x-swagger-router-controller": "Friend",
      "tags": ["Friend"],
      "operationId": "friendAnswerFriendRequestPost",
      "parameters": [
        {
  "name" : "user_id",
  "in" : "query",
  "description" : "شناسه کاربر درخواست دهنده",
  "required" : false,
  "type" : "number"
},
        {
  "name" : "request_id",
  "in" : "query",
  "description" : "شناسه درخواست ارسال شده",
  "required" : false,
  "type" : "number"
},
        {
  "name" : "answer",
  "in" : "query",
  "description" : "پاسخ درخواست به صورت پذیرش یا رد",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/friend/send_friend_request": {
    
    "post": {
      "summary": "send friend request",
      "description":"ارسال درخواست دوستی",
      "x-swagger-router-controller": "Friend",
      "tags": ["Friend"],
      "operationId": "friendSendFriendRequestPost",
      "parameters": [
        {
  "name" : "user_id",
  "in" : "query",
  "description" : "شناسه کاربر درخواست دهنده",
  "required" : false,
  "type" : "string"
},
        {
  "name" : "target_id",
  "in" : "query",
  "description" : "شناسه کاربر هدف",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/friend/search_user": {
    
    "get": {
      "summary": "search in users",
      "description":"جستجو بین کاربران",
      "x-swagger-router-controller": "Friend",
      "tags": ["Friend"],
      "operationId": "friendSearchUserGet",
      "parameters": [
        {
  "name" : "user_id",
  "in" : "query",
  "description" : "شناسه کاربر درخواست دهنده",
  "required" : false,
  "type" : "number"
},
        {
  "name" : "query",
  "in" : "query",
  "description" : "متن جستجو",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "An array of Users",
  "schema" : {
    "type" : "array",
    "items" : {
      "$ref" : "#/definitions/User"
    }
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/friend/block_user": {
    
    "post": {
      "summary": "block a user",
      "description":"بلاک کردن کاربر",
      "x-swagger-router-controller": "Friend",
      "tags": ["Friend"],
      "operationId": "friendBlockUserPost",
      "parameters": [
        {
  "name" : "user_id",
  "in" : "query",
  "description" : "شناسه کاربر درخواست دهنده",
  "required" : false,
  "type" : "number"
},
        {
  "name" : "target_id",
  "in" : "query",
  "description" : "شناسه کاربر هدف",
  "required" : false,
  "type" : "number"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/friend/get_friends_list": {
    
    "get": {
      "summary": "get list of friends",
      "description":"دریافت لیست دوستان شامل اطلاعات اولیه کاربران",
      "x-swagger-router-controller": "Friend",
      "tags": ["Friend"],
      "operationId": "friendGetFriendsListGet",
      "parameters": [
        {
  "name" : "user_id",
  "in" : "query",
  "description" : "نام کاربری شخص درخواست دهنده.",
  "required" : true,
  "type" : "string",
  "format" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "type" : "array",
    "items" : {
      "$ref" : "#/definitions/User"
    }
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/friend/get_user_status": {
    
    "get": {
      "summary": "get user status",
      "description":"دریافت وضعیت انلاین بودن کاربر",
      "x-swagger-router-controller": "Friend",
      "tags": ["Friend"],
      "operationId": "friendGetUserStatusGet",
      "parameters": [
        {
  "name" : "user_id",
  "in" : "query",
  "description" : "شناسه کاربر درخواست دهنده",
  "required" : false,
  "type" : "number"
},
        {
  "name" : "target_id",
  "in" : "query",
  "description" : "شناسه کاربر هدف",
  "required" : false,
  "type" : "number"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/User"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/chat/send_message": {
    
    "post": {
      "summary": "send message",
      "description":"ارسال پیام",
      "x-swagger-router-controller": "Chat",
      "tags": ["Chat"],
      "operationId": "chatSendMessagePost",
      "parameters": [
        {
  "name" : "user_id",
  "in" : "query",
  "description" : "شناسه کاربر ارسال کننده",
  "required" : true,
  "type" : "number"
},
        {
  "name" : "room_id",
  "in" : "query",
  "description" : "گروه هدف",
  "required" : true,
  "type" : "number"
},
        {
  "name" : "photo",
  "in" : "query",
  "description" : "تصویر پیام",
  "required" : true,
  "type" : "number",
  "format" : "string"
},
        {
  "name" : "video",
  "in" : "query",
  "description" : "ویدیو  پیام",
  "required" : true,
  "type" : "number",
  "format" : "string"
},
        {
  "name" : "text",
  "in" : "query",
  "description" : "متن پیام",
  "required" : false,
  "type" : "number"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/friend/add_user_to_room": {
    
    "post": {
      "summary": "add user to room",
      "description":"اضافه کردن یک کاربر به یک گروه",
      "x-swagger-router-controller": "Chat",
      "tags": ["Chat"],
      "operationId": "friendAddUserToRoomPost",
      "parameters": [
        {
  "name" : "user_id",
  "in" : "query",
  "description" : "شناسه کاربر درخواست دهنده",
  "required" : true,
  "type" : "number",
  "format" : "string"
},
        {
  "name" : "target_id",
  "in" : "query",
  "description" : "شناسه کاربر هدف",
  "required" : true,
  "type" : "number",
  "format" : "string"
},
        {
  "name" : "room_id",
  "in" : "query",
  "description" : "شناسه گروه هدف",
  "required" : true,
  "type" : "number",
  "format" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/chat/get_user_chat": {
    
    "get": {
      "summary": "get details of a user chat",
      "description":"اطلاعات چت بین دو کاربر نصبت به آخرین زمان سینک دیتابیس کلاینت گرفته می‌شود",
      "x-swagger-router-controller": "Chat",
      "tags": ["Chat"],
      "operationId": "chatGetUserChatGet",
      "parameters": [
        {
  "name" : "user_id",
  "in" : "query",
  "description" : "شناسه کاربر درخواست دهنده",
  "required" : true,
  "type" : "number",
  "format" : "string"
},
        {
  "name" : "last_update",
  "in" : "query",
  "description" : "last client database update date.",
  "required" : true,
  "type" : "string",
  "format" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "لیست پیام های کاربر.",
  "schema" : {
    "type" : "array",
    "items" : {
      "$ref" : "#/definitions/Message"
    }
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/chat/start_chat_on_a_place": {
    
    "post": {
      "summary": "create a topic on a place",
      "description":"یک گروه برای یک مکان ایجاد می‌کنیم که چندین کاربر قادر به صحبت در آن هستند",
      "x-swagger-router-controller": "Chat",
      "tags": ["Chat"],
      "operationId": "chatStartChatOnAPlacePost",
      "parameters": [
        {
  "name" : "user_id",
  "in" : "query",
  "description" : "شناسه کاربر ارسال کننده",
  "required" : true,
  "type" : "number",
  "format" : "string"
},
        {
  "name" : "plcae_id",
  "in" : "query",
  "description" : "مکان مورد نظر",
  "required" : true,
  "type" : "number",
  "format" : "string"
},
        {
  "name" : "users",
  "in" : "query",
  "description" : "لیستی از کاربران دعوت شده",
  "required" : true,
  "type" : "string",
  "format" : "array"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/chat/get_chat_list": {
    
    "get": {
      "summary": "get list of chats",
      "description":"گرفتن لیست چت‌های کاربر برای نمایش در لیست ویو کلاینت. این لیست بر اساس آخرین بروزرسانی دیتابیس کلاینت صورت می‌گیرد.",
      "x-swagger-router-controller": "Chat",
      "tags": ["Chat"],
      "operationId": "chatGetChatListGet",
      "parameters": [
        {
  "name" : "user_id",
  "in" : "query",
  "description" : "شناسه کاربر درخواست دهنده",
  "required" : true,
  "type" : "number",
  "format" : "string"
},
        {
  "name" : "last_update",
  "in" : "query",
  "description" : "last client database update date.",
  "required" : true,
  "type" : "string",
  "format" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "An array of Messages",
  "schema" : {
    "type" : "array",
    "items" : {
      "$ref" : "#/definitions/Message"
    }
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
...
  }, "definitions": {
  "User": {
  "type" : "object",
  "properties" : {
    "id" : {
      "type" : "string",
      "description" : "id."
    },
    "phone_number" : {
      "type" : "number",
      "description" : "phone number of the user."
    },
    "username" : {
      "type" : "string",
      "description" : "username of user"
    },
    "password" : {
      "type" : "string",
      "description" : "password of user"
    },
    "avatar" : {
      "type" : "string",
      "description" : "Image URL of the user."
    }
  }
},"Message": {
  "type" : "object",
  "properties" : {
    "id" : {
      "type" : "string",
      "description" : "chat id."
    },
    "sender_id" : {
      "type" : "number",
      "description" : "sender id."
    },
    "room_id" : {
      "type" : "number",
      "description" : "room id"
    },
    "text" : {
      "type" : "string",
      "description" : "message text"
    },
    "photo" : {
      "type" : "string",
      "description" : "Image URL of the message."
    },
    "video" : {
      "type" : "string",
      "description" : "Video URL of the user."
    },
    "date" : {
      "type" : "string",
      "description" : "send time."
    },
    "seen" : {
      "type" : "boolean",
      "description" : "user seen this message?"
    }
  }
},"Place": {
...
},"Error": {
  "type" : "object",
  "properties" : {
    "code" : {
      "type" : "integer",
      "format" : "int32"
    },
    "message" : {
      "type" : "string"
    }
  }
},"DoneResponse": {
  "type" : "object",
  "properties" : {
    "code" : {
      "type" : "integer",
      "format" : "int32"
    },
    "message" : {
      "type" : "string"
    }
  }
}
  }
}